解决分配角色时.如果没有用户就异常问题

+++ src/main/java/com/lanyuan/controller/system/ResourcesController.java
@@ -247,6 +247,9 @@ public class ResourcesController extends BaseController {
 			}
 		}
 		userId = Common.trimComma(userId);
+		if(Common.isEmpty(userId)){
+			return "分配失败,该角色下没有用户!";
+		}
 		String[] users = userId.split(",");
 		for (String uid : users) {
 			resourcesMapper.deleteByAttribute("userId", uid, ResUserFormMap.class);
